add justfile and move secrets to submodule
- add justfile with fmt, check, and secrets commands
- add just, nixfmt, stylua to devshell
- replace remote secrets flake input with local submodule
- simplify sops.nix to use submodule path
