+++ home/nixos/desktop/hyprland/hyprland.nix
@@ -84,33 +84,33 @@
       kdePackages.qtsvg
     ];
 
-  #   programs.fuzzel = {
-  #     enable = true;
-  #     settings = {
-  #       main = {
-  #         font = "Gabarito";
-  #         terminal = "${pkgs.kitty}/bin/kitty";
-  #         prompt = "_> ";
-  #         layer = "overlay";
-  #       };
-  #
-  #       colors = {
-  #         background = "000000ff";
-  #         text = "e2e2e2ff";
-  #         selection = "242424ff";
-  #         selection-text = "e2e2e2ff";
-  #         border = "242424ff";
-  #         match = "e2e2e2ff";
-  #         selection-match = "e2e2e2ff";
-  #       };
-  #
-  #       border = {
-  #         radius = 17;
-  #         width = 2;
-  #       };
-  #
-  #       dmenu.exit-immediately-if-empty = "yes";
-  #     };
-  #   };
+    #   programs.fuzzel = {
+    #     enable = true;
+    #     settings = {
+    #       main = {
+    #         font = "Gabarito";
+    #         terminal = "${pkgs.kitty}/bin/kitty";
+    #         prompt = "_> ";
+    #         layer = "overlay";
+    #       };
+    #
+    #       colors = {
+    #         background = "000000ff";
+    #         text = "e2e2e2ff";
+    #         selection = "242424ff";
+    #         selection-text = "e2e2e2ff";
+    #         border = "242424ff";
+    #         match = "e2e2e2ff";
+    #         selection-match = "e2e2e2ff";
+    #       };
+    #
+    #       border = {
+    #         radius = 17;
+    #         width = 2;
+    #       };
+    #
+    #       dmenu.exit-immediately-if-empty = "yes";
+    #     };
+    #   };
   };
 }
