+++ home/nixos/desktop/hyprland/scripts/bitwarden-float.nix
@@ -1,37 +1,37 @@
 pkgs:
 
 pkgs.writeScript "hyprland-bitwarden-resize" ''
-#!/bin/sh
+  #!/bin/sh
 
-timeout=0
-handle() {
-  case $1 in
-    windowtitle*)
-      # Extract the window ID from the line
-      window_id=''${1#*>>}
+  timeout=0
+  handle() {
+    case $1 in
+      windowtitle*)
+        # Extract the window ID from the line
+        window_id=''${1#*>>}
 
-      # Fetch the list of windows and parse it using jq to find the window by its decimal ID
-      window_info=$(hyprctl clients -j | ${pkgs.jq}/bin/jq --arg id "0x$window_id" '.[] | select(.address == ($id))')
+        # Fetch the list of windows and parse it using jq to find the window by its decimal ID
+        window_info=$(hyprctl clients -j | ${pkgs.jq}/bin/jq --arg id "0x$window_id" '.[] | select(.address == ($id))')
 
-      # Extract the title from the window info
-      window_title=$(echo "$window_info" | ${pkgs.jq}/bin/jq '.title')
+        # Extract the title from the window info
+        window_title=$(echo "$window_info" | ${pkgs.jq}/bin/jq '.title')
 
-      # Check if the title matches the characteristics of the Bitwarden popup window
-      if [[ "$window_title" == *"(Bitwarden"*"Password Manager) - Bitwarden"* ]]; then
-      
-        # in case of double rename only allow once per second
-          if ((SECONDS < timeout)); then
-              return
+        # Check if the title matches the characteristics of the Bitwarden popup window
+        if [[ "$window_title" == *"(Bitwarden"*"Password Manager) - Bitwarden"* ]]; then
+        
+          # in case of double rename only allow once per second
+            if ((SECONDS < timeout)); then
+                return
+            fi
+            timeout=$((SECONDS + 1))
+
+            # echo "$window_id", "$window_title"
+            hyprctl --batch "dispatch togglefloating address:0x$window_id; dispatch resizewindowpixel exact 20% 54%,address:0x$window_id; dispatch centerwindow"
           fi
-          timeout=$((SECONDS + 1))
+        ;;
+    esac
+  }
 
-          # echo "$window_id", "$window_title"
-          hyprctl --batch "dispatch togglefloating address:0x$window_id; dispatch resizewindowpixel exact 20% 54%,address:0x$window_id; dispatch centerwindow"
-        fi
-      ;;
-  esac
-}
-
-# Listen to the Hyprland socket for events and process each line with the handle function
-${pkgs.socat}/bin/socat -U - UNIX-CONNECT:"$XDG_RUNTIME_DIR"/hypr/$HYPRLAND_INSTANCE_SIGNATURE/.socket2.sock | while read -r line; do handle "$line"; done
+  # Listen to the Hyprland socket for events and process each line with the handle function
+  ${pkgs.socat}/bin/socat -U - UNIX-CONNECT:"$XDG_RUNTIME_DIR"/hypr/$HYPRLAND_INSTANCE_SIGNATURE/.socket2.sock | while read -r line; do handle "$line"; done
 ''
