+++ home/shared/nvim/lua/config/plugins/treesitter.lua
@@ -2,6 +2,6 @@
 -- Already loaded as a startup plugin, this just adds extra configuration
 
 return {
-  -- nvim-treesitter-textobjects is not in our plugin list yet
-  -- but would be configured here if we add it
+	-- nvim-treesitter-textobjects is not in our plugin list yet
+	-- but would be configured here if we add it
 }
