+++ home/shared/nvim/nix/packages.nix
@@ -7,18 +7,20 @@
 # general.extra, general.telescope, general.blink).
 # You can also enable specific subcategories: `general.always = true`.
 {
-  nvim = { pkgs, ... }: {
-    settings = {
-      wrapRc = true;        # Bake lua config into the nix store (reproducible)
-      configDirName = "nvim"; # NVIM_APPNAME — used for ~/.config/<name> when wrapRc = false
-    };
+  nvim =
+    { pkgs, ... }:
+    {
+      settings = {
+        wrapRc = true; # Bake lua config into the nix store (reproducible)
+        configDirName = "nvim"; # NVIM_APPNAME — used for ~/.config/<name> when wrapRc = false
+      };
 
-    categories = {
-      general = true;       # All general.* subcategories
-      treesitter = true;    # Syntax highlighting
-      lsp = true;           # LSP servers + related plugins
-      themer = true;        # Colorscheme
-      colorscheme = "catppuccin";
+      categories = {
+        general = true; # All general.* subcategories
+        treesitter = true; # Syntax highlighting
+        lsp = true; # LSP servers + related plugins
+        themer = true; # Colorscheme
+        colorscheme = "catppuccin";
+      };
     };
-  };
 }
